+++ catalogo-strumenti-backend/catalogo-backend-service/src/main/java/it/istat/mec/catalog/dao/DocumentationDao.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import it.istat.mec.catalog.domain.Documentation;
 
 @Repository
-public interface DocumentationDao extends JpaRepository<Documentation, Long> {
+public interface DocumentationDao extends JpaRepository<Documentation, Integer> {
 
 	@Override
 	List<Documentation> findAll();
